+++ routes/api.php
@@ -70,6 +70,7 @@ Route::middleware(['web', 'auth'])->group(function () {
     Route::get('/tickets/{ticket}/hours', [TicketController::class, 'getHours']);
     Route::post('/tickets/{ticket}/hours', [TicketController::class, 'logHours']);
     Route::get('/tickets/{ticket}/export-hours', [TicketController::class, 'exportHours']);
+    Route::put('/tickets/{ticket}/dates', [TicketController::class, 'updateDates']);
     
     Route::apiResource('users', UserController::class);